firebase-func/requirements.txt
azure-core==1.18.0
azure-cosmos==4.2.0
azure-functions==1.7.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
cryptography==3.4.8
gcloud==0.17.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
jws==0.1.3
jwt==1.2.0
oauth2client==3.0.0
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pyparsing==2.4.7
Pyrebase==3.0.18
python-jwt==2.0.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.7.0
rsa==4.7.2
six==1.16.0
sseclient==0.0.27
urllib3==1.26.6
